Use (Local)StructuredProperty to collect event meta data.
Now, the meta includes:
 - version (int)
 - created (datetime)

Put created datetime in meta instead of in each PublishedEvent is trying to
reduce the query cost when we add 'whatsnew' function in the future.
If the created datetime is located in each PublishedEvent, we may need to
query all PublishedEvent and filter on created field. If created datetime
is located in manager's meta field, we just need query one manager entity
which is memory cached.